chore: Updated primitive type for move speed method

+++ core/java/net/sf/l2j/gameserver/model/actor/stat/BoatStat.java
@@ -39,7 +39,7 @@ public class BoatStat extends CharStat
     }
 
     @Override
-    public final float getMoveSpeed()
+    public final double getMoveSpeed()
     {
         return _moveSpeed;
     }
